Refactored Modules to Elements, Refactored and made StandardModule to AbstractElement
